Follow the redux book standard
