+++ generators/app/templates/client/src/components/Header/Header.component.js
@@ -1,6 +1,8 @@
+
 import React, { Component } from 'react';
 import { IntlProvider, addLocaleData } from 'react-intl';
 import PropTypes from 'prop-types';
+
 import fr from 'react-intl/locale-data/fr';
 import frMessages from '../../translations/fr.json';
 import logo from './logo.svg';
