+++ generators/app/templates/client/src/components/Header/Header.container.js
@@ -1,0 +1,15 @@
+import Root from './Header.component';
+import { connect } from 'react-redux';
+
+function mapStateToProps(state) {
+  return {};
+}
+
+function mapDispatchToProps(dispatch) {
+  return {};
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(Root);
