+++ generators/app/templates/client/src/components/index.js
@@ -1,0 +1,5 @@
+/**
+ * Index for components
+ */
+
+export { default as Header } from './Header';
