+++ generators/app/templates/client/src/containers/Home/Home.container.js
@@ -1,0 +1,15 @@
+import { connect } from 'react-redux';
+import Home from './Home.component';
+
+function mapStateToProps(state) {
+  return {};
+}
+
+function mapDispatchToProps(dispatch) {
+  return {};
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(Home);
