+++ generators/app/templates/client/src/containers/Home/index.js
@@ -1,0 +1,2 @@
+export { default } from './Home.container';
+
