+++ generators/app/templates/client/src/containers/Home/style.css
@@ -2,6 +2,6 @@
   text-align: center;
 }
 
-.App-intro {
+.intro {
   font-size: large;
 }
