+++ generators/app/templates/client/src/containers/Page/style.css
