+++ generators/app/templates/client/src/containers/index.js
@@ -1,0 +1,6 @@
+/**
+ * Index for containers
+ */
+
+export { default as Home } from './Home';
+export { default as Page } from './Page';
