--- generators/app/templates/client/src/containers/root/index.js
@@ -1,1 +1,0 @@
-export { default } from './root.container';
